+++ src/napi.rs
@@ -1,12 +1,5 @@
 #[link(name = "ace_napi.z")]
 extern "C" {}
 
-mod napi_api10;
-pub use napi_api10::*;
-
-#[cfg(feature = "api-11")]
-#[cfg_attr(docsrs, doc(cfg(feature = "api-11")))]
-mod api11_additions;
-#[cfg(feature = "api-11")]
-#[cfg_attr(docsrs, doc(cfg(feature = "api-11")))]
-pub use api11_additions::*;
+mod napi_ffi;
+pub use napi_ffi::*;